Move files under src
